backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
websockets = "^15.0.1"
python-socketio = "^5.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
